consolidated state withholding logic to 1 file
