+++ src/utils/index.ts
@@ -1,6 +1,5 @@
 // file that allows you to just import from /utils
 export { social_security_withholding, medicare_withholding, federal_withholding } from './federal_withholding';
-export { determineStateTaxesWithheld } from './state_withholding';
-export { US_STATES_MAP, instanceOfTaxUnknown } from './us_states';
+export { US_STATES_MAP, instanceOfTaxUnknown, determineStateTaxesWithheld } from './state_withholding';
 
 export { formatCurrency } from './helperFunctions';
